+++ packages/connect-ucan/src/connect-ucan-account.ts
@@ -5,8 +5,8 @@ import { delegate, Delegation } from '@ucanto/core'
 import { fireproof } from '@fireproof/core'
 import type { Doc, Database, MapFn } from '@fireproof/core'
 
-import { ConnectIPFSParams } from './connect-ipfs'
-import { DatabaseConnectIPFS } from './connect-ipfs-helpers'
+import { ConnectIPFSParams } from './connect-ucan'
+import { DatabaseConnectIPFS } from './connect-ucan-helpers'
 import { Capabilities } from '@ucanto/interface'
 // import { Loader } from '@fireproof/encrypted-blockstore'
 import { OwnedSpace, Space } from '@web3-storage/w3up-client/space'
